02-sk-orchestration/pyproject.toml
[project]
name = "mcp-demo"
version = "0.1.0"
description = ""
authors = [
    {name = "Frida Almogabar",email = "falmogabarsanchez@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "mcp (>=1.10.1,<2.0.0)",
    "fastmcp (>=2.10.2,<3.0.0)",
    "azure-ai-ml (>=1.27.1,<2.0.0)",
    "azure-ai-projects (>=1.0.0b12,<2.0.0)",
    "azure-ai-agents (== 1.1.0b3 )",
    "langchain-azure-ai (>=0.1.4,<0.2.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "semantic-kernel (>=1.34.0,<2.0.0)",
    "fastapi (>=0.104.0,<1.0.0)",
    "uvicorn (>=0.24.0,<1.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "httpx (>=0.25.0,<1.0.0)",
    "pydantic (>=2.0.0,<3.0.0)",
    "starlette (>=0.28.0,<1.0.0)"
]



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
